Calendar – Start at today + Clickable Sidewidget + Breadcrumbs, closed #81 - added the today button - added scroll to today when page was loaded - fixed the breadcrumb logic for more then 2 items - updated nested meeting components for avoid to confusing
