+++ resources/views/meeting-list.blade.php
@@ -16,85 +16,41 @@ if (isset($meetings) and is_array($meetings)) {
 ?>
 
 @section('content')
-    <main class="ris-main ris-calendar">
+    <meeting-list inline-template>
 
-        @include('layouts.breadcrumbs')
+        <main class="ris-main ris-calendar">
 
-        <h1 class="ris-h3 ris-calendar__headline">
-            Sitzungskalender
-        </h1>
+            @include('layouts.breadcrumbs')
 
-        <div class="ris-calendar__flex-wrapper ris-content ris-content_has-widget ris-content_six-eight-eight">
-            <calendar-plugin></calendar-plugin>
+            <h1 class="ris-h3 ris-calendar__headline">
+                Sitzungskalender
+                <button @click.prevent="goToToday" class="ris-button ris-button_secondary" title="Heute">
+                    <span class="ris-i ris-i_calendar"></span>Heute
+                </button>
+            </h1>
+            <div class="ris-calendar__flex-wrapper ris-content ris-content_has-widget ris-content_six-eight-eight">
+                <calendar-plugin></calendar-plugin>
+                <div class="ris-calendar__content">
+                    <div class="ris-calendar__card-list-wrapper" data-page-loaded="1" id="card-list">
 
-            <div class="ris-calendar__content">
-                <div class="ris-calendar__card-list-wrapper" data-page-loaded="1" id="card-list">
-
-                    @foreach ($meeting_list_sorted as $meeting_list_per_week)
-
-                        <div class="ris-subheader ris-calendar__card-list-ris-subheader">Kalenderwoche
-                            {{ \Illuminate\Support\Carbon::parse(head($meeting_list_per_week)[0]->dateFrom)->weekOfYear }}
-                        </div>
-
-                        @foreach ($meeting_list_per_week as $meeting_list_per_day)
-                            <div class="ris-calendar__card-list">
-                                <section class="ris-calendar__card-day">
-
-                                    <div class="ris-calendar__card-day-left"
-                                        id="{{ '_' . \Illuminate\Support\Carbon::parse($meeting_list_per_day[0]->dateFrom)->day
-                                        . '_' . \Illuminate\Support\Carbon::parse($meeting_list_per_day[0]->dateFrom)->month
-                                        . '_' . \Illuminate\Support\Carbon::parse($meeting_list_per_day[0]->dateFrom)->year }}"
-                                        >
-                                        {{ \Illuminate\Support\Carbon::parse($meeting_list_per_day[0]->dateFrom)->day }}
-                                        <br/>
-                                        <span class="ris-calendar__card-day-of-week">
-                                            {{ \Illuminate\Support\Carbon::parse($meeting_list_per_day[0]->dateFrom)->locale('de')->minDayName }}
-                                        </span>
-                                    </div>
-
-                                    <div class="ris-calendar__card-day-right">
-                                        @foreach ($meeting_list_per_day as $meeting)
-                                            <a class="ris-link ris-calendar__card" title="{{ $meeting->title }}"
-                                                href="/meeting/{{ $meeting->id }}"
-                                            >
-                                                <h2 class="ris-title">
-                                                    {{ $meeting->title }}
-                                                </h2>
-                                                <div class="ris-subheader">
-                                                    {{ \Illuminate\Support\Carbon::parse($meeting->dateFrom)->isoFormat('LLLL') }}-{{ \Illuminate\Support\Carbon::parse($meeting->dateTill)->format('H:i') }}
-                                                    Uhr
-                                                </div>
-                                                <div class="ris-session-count">
-                                                    <div class="ris-session-count__agenda">
-                                                        {{ $meeting->agendaCount }}
-                                                        <span class="ris-i ris-i_list"></span>
-                                                    </div>
-                                                    <div class="ris-session-count__people">
-                                                        {{ $meeting->peopleCount }}
-                                                        <span class="ris-i ris-i_people"></span>
-                                                    </div>
-                                                    <div class="ris-session-count__file">
-                                                        {{ $meeting->fileCount }}
-                                                        <span class="ris-i ris-i_download"></span>
-                                                    </div>
-                                                </div>
-                                            </a>
-                                        @endforeach
-                                    </div>
-
-                                </section>
+                        @foreach ($meeting_list_sorted as $meeting_list_per_week)
+                            <div class="ris-subheader ris-calendar__card-list-ris-subheader">Kalenderwoche
+                                {{ \Illuminate\Support\Carbon::parse(head($meeting_list_per_week)[0]->dateFrom)->weekOfYear }}
                             </div>
+
+                            @include('components.meeting.meeting-list',
+                                ['meeting_list' => $meeting_list_per_week]
+                            )
                         @endforeach
 
-                    @endforeach
-
-                    <meeting-list></meeting-list>
+                        <meeting-list-lazy></meeting-list-lazy>
+                    </div>
                 </div>
-
             </div>
 
-        </div>
-    </main>
+        </main>
+
+    </meeting-list>
 
     @include('layouts.footer')
